wasm2c: Always use a module prefix and set one by default
The module prefix is no longer optional and is now set by default to the
name of the module as given in the name section, or taken from the name
of the input file.

A new `-n/--module-name` command line flag can also be used to override.

The prefix used by the generated code is now fixed at wasm2c time which
makes the output code easier to read and also avoid the symbol pasting
in the C pre-processor which makes the source hard to understand.  For
example, it makes symbols hard to `grep` for.